+++ controller/topup_waffo_pancake.go
@@ -160,7 +160,7 @@ func RequestWaffoPancakePay(c *gin.Context) {
 	tradeNo := fmt.Sprintf("WAFFO_PANCAKE-%d-%d-%s", id, time.Now().UnixMilli(), randstr.String(6))
 	topUp := &model.TopUp{
 		UserId:        id,
-		Amount:        normalizeWaffoPancakeTopUpAmount(req.Amount),
+		Amount:        float64(normalizeWaffoPancakeTopUpAmount(req.Amount)),
 		Money:         payMoney,
 		TradeNo:       tradeNo,
 		PaymentMethod: model.PaymentMethodWaffoPancake,
